:recycle: (client/DecodePassword) Refactor Decode page into other component
