+++ src/components/decode/ErrorMessage.tsx
@@ -1,0 +1,13 @@
+interface ErrorMessageProps {
+  error: string | null;
+}
+
+export default function ErrorMessage({ error }: ErrorMessageProps) {
+  if (!error) return null;
+  
+  return (
+    <div className="mb-4 p-3 bg-red-50 dark:bg-red-900/20 border border-red-200 dark:border-red-800 rounded-lg text-red-600 dark:text-red-400 text-sm">
+      {error}
+    </div>
+  );
+}
